refactor: extract flushUnused
